Ignore discord updates when the message hasn't been edited in the last 6 hours
Discord sends spurious updates for messages that haven't been edited, or
haven't been edited recently. Messages from as far back as 5 years ago have
been observed to be updated, always with a URL or embedded content.

These need to be ignored because it's confusing when out of date messages
are bridged as new messages with no context to protocols that don't support
edits. Edited messages normally arrive as updates immediately.